PYTHON-913 - Suppress read preference warning under MongoClient.
